requirements.txt
Django==5.1.6
django-debug-toolbar==5.0.1
django-extensions==3.2.3
ipdb==0.13.13
ipython==9.0.1
isort==6.0.0
psycopg==3.2.4
pytest==8.3.4
pytest-django==4.9.0
pytest-watch==4.2.0
black==25.1.0
flake8==7.1.1
